Add fetcher prototype
