+++ oap-server/server-bootstrap/src/main/resources/fetcher-prom-rules/self.yaml
@@ -1,0 +1,66 @@
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# Only for test.
+fetcherInterval: PT15S
+fetcherTimeout: PT10S
+metricsPath: /metrics
+staticConfig:
+  # targets will be labeled as "instance"
+  targets:
+    - localhost:1234
+  labels:
+    app: test-oap
+metricsRules:
+  - name: instance_cpu_percentage
+    scope: SERVICE_INSTANCE
+    downSampling: doubleAvg
+    sources:
+      process_cpu_seconds_total:
+        counterFunction: RATE
+        range: PT1M
+        relabel:
+          service:
+            - app
+          instance:
+            - instance
+  - name: service_cpu_percentage
+    scope: SERVICE
+    downSampling: doubleAvg
+    sources:
+      process_cpu_seconds_total:
+        counterFunction: RATE
+        range: PT1M
+        relabel:
+          service:
+            - app
+  - name: instance_jvm_memory_bytes_used
+    scope: SERVICE_INSTANCE
+    downSampling: doubleAvg
+    sources:
+      jvm_memory_bytes_used:
+        relabel:
+          service:
+            - app
+          instance:
+            - instance
+  - name: service_jvm_memory_bytes_used
+    scope: SERVICE
+    downSampling: doubleAvg
+    sources:
+      jvm_memory_bytes_used:
+        relabel:
+          service:
+            - app
