+++ oap-server/server-core/pom.xml
@@ -78,10 +78,6 @@
             <groupId>org.javassist</groupId>
             <artifactId>javassist</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-text</artifactId>
-        </dependency>
 
         <dependency>
             <groupId>org.apache.skywalking</groupId>
