+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/metrics/DoubleAvgMetrics.java
@@ -37,11 +37,11 @@ public abstract class DoubleAvgMetrics extends Metrics implements DoubleValueHol
     @Getter
     @Setter
     @Column(columnName = SUMMATION, storageOnly = true)
-    private double summation;
+    protected double summation;
     @Getter
     @Setter
     @Column(columnName = COUNT, storageOnly = true)
-    private long count;
+    protected long count;
     @Getter
     @Setter
     @Column(columnName = VALUE, dataType = Column.ValueDataType.COMMON_VALUE, function = Function.Avg)
