+++ oap-server/server-fetcher-plugin/prometheus-fetcher-plugin/pom.xml
@@ -35,5 +35,13 @@
             <artifactId>server-core</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.yaml</groupId>
+            <artifactId>snakeyaml</artifactId>
+        </dependency>
+         <dependency>
+            <groupId>com.squareup.okhttp3</groupId>
+            <artifactId>okhttp</artifactId>
+         </dependency>
     </dependencies>
 </project>
